+++ backend/mail-server/src/main/java/news/bombomemail/common/config/TimeConfig.java
@@ -1,4 +1,4 @@
-package news.bombomemail.common;
+package news.bombomemail.common.config;
 
 import java.time.Clock;
 import java.time.ZoneId;
